broken code, but the deadline is close, time to make video

+++ backend/src/main/java/no/ntnu/exam2021/backend/cart/CartRepository.java
@@ -7,8 +7,10 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CartRepository extends JpaRepository<Cart, Integer> {
+public interface CartRepository extends JpaRepository<Cart, Long> {
     List<Cart> findAllByUserOrderByCreatedDateDesc(User user);
 
     List<Cart> deleteByUser(User user);
+
+
 }
