+++ backend/src/main/java/no/ntnu/exam2021/backend/cart/CartService.java
@@ -57,7 +57,7 @@ public class CartService {
         cartRepository.save(cart);
     }
 
-    public void deleteCartItem(int id,int userId) throws CartItemNotExistException {
+    public void deleteCartItem(Long id,Long userId) throws CartItemNotExistException {
         if (!cartRepository.existsById(id))
             throw new CartItemNotExistException("Cart id is invalid : " + id);
         cartRepository.deleteById(id);
